refactor: move KanjiMark into commons to fix inverted dependency
`KanjiMark` was declared in the background listener that happens to produce it,
so `commons/message.ts` and `commons/addFurigana.ts` -- the shared lower layer --
imported from `entrypoints/background/listeners/`, the upper layer.

`KanjiMark` describes the payload of the `getKanjiMarks` message, not an
implementation detail of the listener, so `commons/message.ts` next to the
`ProtocolMap` that uses it is its natural home. The background listener and the
Playground component now import it from there like every other consumer, and
`commons` no longer imports from `entrypoints` at all.

Type-only move, no behaviour change.
