+++ extension/src/commons/message.ts
@@ -3,7 +3,16 @@ import { defineExtensionMessaging } from "@webext-core/messaging";
 export type { ExtensionMessage } from "@webext-core/messaging";
 
 import type { FuriganaType } from "@/commons/constants";
-import type { KanjiMark } from "@/entrypoints/background/listeners/onGetKanjiMarksMessage";
+import type { KanjiToken } from "@/commons/toKanjiToken";
+
+/**
+ * A {@link KanjiToken} as it crosses the wire, tagged with whether the user's kanji
+ * filter rules match it. Lives here rather than in the background listener so that
+ * `commons` never has to depend on `entrypoints`.
+ */
+export interface KanjiMark extends KanjiToken {
+  isFiltered: boolean;
+}
 
 interface ProtocolMap {
   addFurigana(): void;
